+++ ObjectPosition.h
@@ -17,8 +17,8 @@ public:
 	double robotAngle;
 	void setFrameSize(cv::Size frameSize);
 
-	virtual void updateCoordinates(int x, int y, cv::Point robotFieldCoords, int robotAngle); // Takes raw coordinates of object from frame
-	virtual void updateCoordinates(cv::Point point, cv::Point robotFieldCoords, int robotAngle);
+	virtual void updateCoordinates(int x, int y, cv::Point robotFieldCoords, double robotAngle); // Takes raw coordinates of object from frame
+	virtual void updateCoordinates(cv::Point point, cv::Point robotFieldCoords, double robotAngle);
 	virtual void updatePolarCoords(int x, int y);
 	virtual void updatePolarCoords(cv::Point point);
